Created a solver to solve any game that has one solution.
